Decoupled redis interface and engine loop
